Give GetDB it's own  file
